% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_data_fetch.R
\name{coloc_fetch_metadata}
\alias{coloc_fetch_metadata}
\title{Fetch exposure metadata (N, type, ncase/ncontrol) from OpenGWAS}
\usage{
coloc_fetch_metadata(
  exposure_id,
  cache_dir = ardmr_cache_dir(),
  verbose = TRUE
)
}
\description{
Fetch exposure metadata (N, type, ncase/ncontrol) from OpenGWAS
}
\keyword{internal}
