% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_business_logic.R
\name{.choose_weights}
\alias{.choose_weights}
\title{Choose weights ("none" or "inv_se2")}
\usage{
.choose_weights(se, scheme = c("inv_se2", "none"))
}
\description{
Choose weights ("none" or "inv_se2")
}
\keyword{internal}
