% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_business_logic.R
\name{.perm_test_signed}
\alias{.perm_test_signed}
\title{Core permutation engine for signed sums (two-sided p)}
\usage{
.perm_test_signed(
  contrib_signed,
  y,
  exact_max_combn = 1e+05,
  mc_B = 10000,
  seed = NULL,
  retain_perm = FALSE,
  retain_perm_max = 10000
)
}
\description{
Core permutation engine for signed sums (two-sided p)
}
\keyword{internal}
