% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_business_logic.R
\name{.perm_test_sum}
\alias{.perm_test_sum}
\title{Core permutation engine for non-negative tail sums (one-sided p)}
\usage{
.perm_test_sum(contrib, y, exact_max_combn = 1e+05, mc_B = 10000, seed = NULL)
}
\description{
Core permutation engine for non-negative tail sums (one-sided p)
}
\keyword{internal}
