% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{get_pkg_obj}
\alias{get_pkg_obj}
\title{Internal: fetch a dataset shipped in the package}
\usage{
get_pkg_obj(name, pkg = "ardmr")
}
\arguments{
\item{name}{String, name of the dataset (without quotes).}

\item{pkg}{Package name (default = "ardmr").}
}
\value{
A tibble with the dataset contents.
}
\description{
Internal: fetch a dataset shipped in the package
}
\keyword{internal}
