% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Manhattan plot over outcomes}
\usage{
manhattan_plot(results_df, Multiple_testing_correction = c("BH", "bonferroni"))
}
\arguments{
\item{results_df}{tidy results}

\item{Multiple_testing_correction}{"BH" or "bonferroni"}
}
\description{
Manhattan plot over outcomes
}
