% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_neale.R
\name{neale_tbi_maker}
\alias{neale_tbi_maker}
\title{Generate .tbi for all bgz files (no-op if present)}
\usage{
neale_tbi_maker(neale_dir, verbose = TRUE)
}
\arguments{
\item{neale_dir}{directory}

\item{verbose}{logical}
}
\description{
Generate .tbi for all bgz files (no-op if present)
}
