% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_beta_mean_cause_compare}
\alias{plot_beta_mean_cause_compare}
\title{plots mean beta by cause for ard_compare}
\usage{
plot_beta_mean_cause_compare(
  beta_cause_tbl,
  title = NULL,
  subtitle = NULL,
  exposure_units,
  effect_scale = c("odds_ratio", "absolute_risk")
)
}
\arguments{
\item{exposure_units}{Character string describing the exposure units used
for the β-scale axis label.}

\item{effect_scale}{Character string specifying whether the β-axis should be
expressed as odds ratios ("odds_ratio") or absolute risk
("absolute_risk").}
}
\description{
plots mean beta by cause for ard_compare
}
\keyword{internal}
