% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_beta_mean_global}
\alias{plot_beta_mean_global}
\title{Global forest plot of IVW mean MR beta (all diseases vs ARDs)}
\usage{
plot_beta_mean_global(
  beta_global_tbl,
  title = NULL,
  exposure_units,
  effect_scale = c("odds_ratio", "absolute_risk")
)
}
\arguments{
\item{beta_global_tbl}{Tibble from beta_mean_global().}

\item{title}{Optional title.}

\item{exposure_units}{Character string describing the exposure units used
for the β-scale axis label.}

\item{effect_scale}{Character string specifying whether the β-axis should be
expressed as odds ratios ("odds_ratio") or absolute risk
("absolute_risk").}
}
\description{
Global forest plot of IVW mean MR beta (all diseases vs ARDs)
}
