% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_beta_mean_global_compare}
\alias{plot_beta_mean_global_compare}
\title{plots the mean beta hypothesis tests globally}
\usage{
plot_beta_mean_global_compare(
  beta_global_tbl,
  title = NULL,
  exposure_units,
  effect_scale = c("odds_ratio", "absolute_risk")
)
}
\arguments{
\item{exposure_units}{Character string describing the exposure units used
for the β-scale axis label.}
}
\description{
plots the mean beta hypothesis tests globally
}
\keyword{internal}
