% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_vcf_reader.R
\name{read_gwasvcf_metadata}
\alias{read_gwasvcf_metadata}
\title{Read aggregate metadata (N, ncase, type) from a GWAS-VCF.}
\usage{
read_gwasvcf_metadata(vcf_path, genome_build = "GRCh37")
}
\description{
Read aggregate metadata (N, ncase, type) from a GWAS-VCF.
}
