% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_vcf_reader.R
\name{read_gwasvcf_region}
\alias{read_gwasvcf_region}
\title{Read a GWAS-VCF region and return a canonical exposure-region tibble.}
\usage{
read_gwasvcf_region(
  vcf_path,
  chr,
  start,
  end,
  genome_build = "GRCh37",
  drop_indels = TRUE,
  info_min = 0.8,
  verbose = TRUE
)
}
\description{
Read a GWAS-VCF region and return a canonical exposure-region tibble.
}
