% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_business_logic.R
\name{signed_ivw_score}
\alias{signed_ivw_score}
\title{Signed IVW evidence: s = sign(beta) * (-log10 p)}
\usage{
signed_ivw_score(beta, p)
}
\description{
Signed IVW evidence: s = sign(beta) * (-log10 p)
}
\keyword{internal}
