% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_vcf_reader.R
\name{validate_gwasvcf}
\alias{validate_gwasvcf}
\title{Validate a GWAS-VCF for use as the coloc exposure source.}
\usage{
validate_gwasvcf(vcf_path, genome_build = "GRCh37")
}
\description{
Validate a GWAS-VCF for use as the coloc exposure source.
}
