% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variant_manifest_downloader.R
\name{Variant_manifest_downloader}
\alias{Variant_manifest_downloader}
\title{Download (if needed) the provider variant manifest (GRCh37) for RSID -> chr:pos mapping}
\usage{
Variant_manifest_downloader(catalog, data_root = "data", overwrite = FALSE)
}
\arguments{
\item{catalog}{Character. One of \code{"neale"}, \code{"panukb"}.}

\item{data_root}{Character. Root directory to store data (default \code{"data"}).}

\item{overwrite}{Logical. If \code{TRUE}, re-download even if the file exists.}
}
\value{
(Invisibly) the local file path to the downloaded (or existing) manifest.
}
\description{
Downloads and caches the large \strong{variant manifest} required to map RSIDs to
genomic positions for a given catalog (\code{"neale"} or \code{"panukb"}). Files are saved under:
\itemize{
\item \code{data/neale/variants.tsv.bgz}
\item \code{data/panukb/full_variant_qc_metrics.txt.bgz}
}

If the file already exists, nothing is downloaded.

Uses \strong{curl} to show a console progress bar. Make sure you have internet access and
enough disk space (~hundreds of MB).
}
