% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano plot over outcomes}
\usage{
volcano_plot(results_df, Multiple_testing_correction = c("BH", "bonferroni"))
}
\arguments{
\item{results_df}{tidy results}

\item{Multiple_testing_correction}{"BH" or "bonferroni"}
}
\description{
Volcano plot over outcomes
}
