% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_snp_mapper.R
\name{exposure_snp_mapper}
\alias{exposure_snp_mapper}
\title{Map RSIDs to provider-specific positions and append columns}
\usage{
exposure_snp_mapper(exposure_snps, sex, cache_dir, verbose = TRUE)
}
\arguments{
\item{exposure_snps}{data frame with rsid}

\item{sex}{"both" -> panukb mapping; otherwise neale mapping}

\item{cache_dir}{path for crosswalk caches}

\item{verbose}{logical}
}
\value{
exposure_snps with new col "panukb_pos" or "neale_pos"
}
\description{
Map RSIDs to provider-specific positions and append columns
}
