% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provider_neale.R
\name{neale_gwas_checker}
\alias{neale_gwas_checker}
\title{Ensure Neale sumstats exist locally (download if needed)}
\usage{
neale_gwas_checker(MR_df, neale_dir, verbose = TRUE)
}
\arguments{
\item{MR_df}{outcomes plan (neale rows)}

\item{neale_dir}{directory with .bgz sumstats}

\item{verbose}{logical}
}
\description{
Ensure Neale sumstats exist locally (download if needed)
}
